+++ platform/jvm/core/src/jni.rs
@@ -1102,6 +1102,23 @@ pub extern "system" fn Java_io_bitdrift_capture_CaptureJniLibrary_setEntityId(
 }
 
 #[no_mangle]
+pub extern "system" fn Java_io_bitdrift_capture_CaptureJniLibrary_clearEntityId(
+  _env: JNIEnv<'_>,
+  _class: JClass<'_>,
+  logger_id: jlong,
+) {
+  with_handle_unexpected(
+    || -> anyhow::Result<()> {
+      let logger = unsafe { LoggerId::from_raw(logger_id) };
+      logger.register_opaque_entity_id(None);
+
+      Ok(())
+    },
+    "jni clear entity id",
+  );
+}
+
+#[no_mangle]
 // Java types are always signed, but log level/type are both unsigned.
 #[allow(clippy::cast_sign_loss)]
 pub extern "system" fn Java_io_bitdrift_capture_CaptureJniLibrary_writeLog(