+++ platform/jvm/gradle-test-app/src/main/java/io/bitdrift/gradletestapp/data/model/Actions.kt
@@ -30,6 +30,8 @@ sealed class ConfigAction : AppAction {
     data class SetSleepModeEnabled(
         val enabled: Boolean,
     ) : ConfigAction()
+
+    object ClearEntityId : ConfigAction()
 }
 
 sealed class SessionAction : AppAction {