+++ platform/jvm/gradle-test-app/src/main/java/io/bitdrift/gradletestapp/data/model/ConfigState.kt
@@ -17,4 +17,5 @@ data class ConfigState(
     val isDeferredStart: Boolean = false,
     val selectedLogLevel: LogLevel = LogLevel.INFO,
     val isSleepModeEnabled: Boolean = false,
+    val entityId: String = "",
 )